+++ enclave-tls/src/tls_wrappers/nulltls/transmit.c
@@ -6,6 +6,7 @@
 #include <unistd.h>
 #include <enclave-tls/log.h>
 #include <enclave-tls/tls_wrapper.h>
+#include <internal/core.h>
 #ifdef SGX
 #include "etls_t.h"
 #endif
@@ -21,7 +22,7 @@ tls_wrapper_err_t nulltls_transmit(tls_wrapper_ctx_t *ctx, void *buf, size_t *bu
 		return TLS_WRAPPER_ERR_TRANSMIT;
 	}
 
-	*buf_size = rc;
+	*buf_size = (size_t)rc;
 
 	return TLS_WRAPPER_ERR_NONE;
 }
